+++ src/components/JobResults/JobListings.vue
@@ -38,7 +38,7 @@
 
 <script>
 import { mapActions, mapGetters } from 'vuex';
-import { FETCH_JOBS, FILTER_JOBS_BY_ORGANIZATIONS } from '@/store';
+import { FETCH_JOBS, FILTER_JOBS_BY_ORGANIZATIONS } from '@/store/constants';
 import JobListing from './JobListing.vue';
 
 export default {
