+++ src/store/constants.js
@@ -1,0 +1,11 @@
+/* USER */
+export const LOGIN_USER = 'LOGIN_USER';
+
+/* JOBS */
+export const RECEIVE_JOBS = 'RECEIVE_JOBS';
+export const FETCH_JOBS = 'FETCH_JOBS';
+
+/* ORGANIZATIONS */
+export const UNIQUE_ORGANIZATIONS = 'UNIQUE_ORGANIZATIONS';
+export const ADD_SELECTED_ORGANIZATIONS = 'ADD_SELECTED_ORGANIZATIONS';
+export const FILTER_JOBS_BY_ORGANIZATIONS = 'FILTER_JOBS_BY_ORGANIZATIONS';
