+++ src/store/state.js
@@ -1,0 +1,7 @@
+const state = () => ({
+  isLoggedIn: false,
+  jobs: [],
+  selectedOrganizations: [],
+});
+
+export default state;
