+++ tests/unit/store/actions.test.js
@@ -1,0 +1,27 @@
+import actions from '@/store/actions';
+import getJobs from '@/api/getJobs';
+
+jest.mock('@/api/getJobs');
+
+describe('actions', () => {
+  describe('FETCH_JOBS', () => {
+    beforeEach(() => {
+      getJobs.mockResolvedValue([{ id: 1, title: 'Software Developer' }]);
+    });
+
+    it('should make a request to fetch jobs', async () => {
+      const context = { commit: jest.fn() };
+      await actions.FETCH_JOBS(context);
+      expect(getJobs).toHaveBeenCalled();
+    });
+
+    it('should send message to save recieved jobs in store', async () => {
+      const commit = jest.fn();
+      const context = { commit };
+      await actions.FETCH_JOBS(context);
+      expect(commit).toHaveBeenCalledWith('RECEIVE_JOBS', [
+        { id: 1, title: 'Software Developer' },
+      ]);
+    });
+  });
+});
